Updated action functions so applyAction now only applies actions.
Consequence actions are now generated in the same functions which generate those actions in the first place, like how moveUnit now also generates all the consequences of moving a unit. The role of applyAction is now to just apply the given action, with no regard to the consequences of the action. Consequent generation should be handled elsewhere.