Several bugfixes and QoL improvements
Fixed an issue where building a unit on top of an enemy unit would merge them.

Fixed an issue where you could only move a unit to capture another unit rather than directly build on top of the enemy unit if the defense power = attack power.

Fixed an issue where capitals were generated for 1 tile provinces.

Fixed an issue where provinces could have tiles in their tile list which did not belong to them.

Fixed an issue where provinces could own tiles but those tiles would not be in their tile list.

Added an income print to the selected province summary print.

Added an update before turn method to better match antiyoy unit starvation rules.

Fixed an issue where building a unit which was immediately merged costed as much as building the resultant merged unit.

Fixed an issue where single tile provinces would not have their structures deleted.

Fixed an issue where undoing an action with a different province than the one who was selected at the time of the action would result in that different province recieving the refund rather than the province actually involved in the action.

Fixed an issue where the capital would not be replaced if the capital tile was the one removed to cause a split.