Bugfixes
Fixed an issue where a faction with no name would cause a crash in printMap.

Fixed an issue where sometime a province would have a new capital generated even if it already had a capital if it was not the biggest contiguous group during a split.

Fixed a very similar issue where if a province of 3 tiles was split into two single tile provinces by losing its capital tile,
one of the new single tile provinces would improperly get a capital placed on it.

Fixed an issue where some structures would not be deleted in single tile provinces, and likewise fixed an issue where capitals would not become trees in single tile provinces.

Fixed an issue where getBuildableUnitsOnTile would not take into account whether or not the building province was active, theoretically allowing an inactive province to build units. Also fixed a similar issue with respect to moving units from inactive province tiles.