Implemented replays
Replays are recordings of a full game, taking a viewer from the initial scenario's state through all the turns until the end of the game. Replays can be saved and loaded to replay files which use JSON representations of everything needed to keep track of what goes on in a game.