Fixed numerous issues with the mark 3.
Ensured that after planInitialAttacks changed things, upgradeNeeds and reserveTiles would be properly recomputed.

Fixed an issue where a unit could be in multiple upgradeNeeds tiers at the same time, thereby causing it to get over-merged.

Fixed an issue where a unit that was merged and ready to attack in the PLAN_CANNIBALIZE_MERGES stage could be doubly cannibalized rather than be allowed to carry out its attack.

Fixed an issue where the upgradeNeeds dictionary used by the PLAN_CANNIBALIZE_MERGES could be altered from its state in PLAN_RESERVIST_MERGES (and likewise for PLAN_UPGRADE_LEFTOVER_UNITS and PLAN_CANNIBALIZE_MERGES).

Added a new state to allow the mark 3 to continue building units even if there are no trees or unclaimed tiles since previously if the mark 3 was surrounded by enemies it could no longer recruit new soldiers unless a tree grew on one of its tiles.

Fixed an issue where planMoveTowardsUnclaimedOrTreesOrEnemies could try to move a unit onto a tile not controlled by the province doing the unit movement.

Updated getMoveTowardsTargetTileAvoidingGivenTiles to use a lambda function to test if a tile should be avoided for efficiency

Fixed an issue in planTreeAndUnclaimedMoves where the same unit could be ordered to both cut down a tree and claim an unclaimed tile, thereby causing an exception.

Fixed another issue in planMoveTowardsUnclaimedOrTreesOrEnemies where tilesToAvoid was not properly updated when a unit moved, thereby making it possible that accidental merges could still happen, including illegal merges that would throw an exception.

Fixed an issue in planCannibalizeMerges where we would try to get the faction of a tile with no owner, causing an exception.

Fixed an issue in both planReservistMerges and planCannibalizeMerges where a unit could be marked as needing an upgrade from a tier that would be too high, like an S3 merging with an S2.